reformat cleanstationame to utils and imported where needed
